requirements.txt
Django==2.2.9
django-markdownx==3.0.1
gunicorn==20.0.4
Markdown==3.1.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
